+++ CodePiece/URLScheme/URLSchemeManager.swift
@@ -14,7 +14,12 @@ final class URLSchemeManager {
 	let eventClass = AEEventClass(kInternetEventClass)
 	let eventID = AEEventID(kAEGetURL)
 	
-	let schemes: [URLScheme.Type] = [ GistScheme.self, SwifterScheme.self ]
+	let schemes: [URLScheme.Type] = [
+		
+		GistScheme.self,
+		SwifterScheme.self,
+		CodePieceScheme.self
+	]
 	
 	init() {
 		eventManager.setEventHandler(self, andSelector: #selector(URLSchemeManager.handleURLEvent(event:withReply:)), forEventClass: eventClass, andEventID: eventID)
@@ -29,7 +34,10 @@ final class URLSchemeManager {
 		
 		if let url = event.url {
 			
-			schemes.forEach { $0.action(url: url) }
+			for scheme in schemes where scheme.match(url: url) {
+
+				scheme.action(url: url)
+			}
 		}
 		else {