+++ CodePiece/Windows/Timeline/ContentsController/MyTweetsContentsController.swift
@@ -18,7 +18,7 @@ final class MyTweetsContentsController : TimelineContentsController, Notificatio
 		return .myTweets
 	}
 	
-	var dataSource = DataSource()
+	var dataSource = ManagedByTweetContentsDataSource()
 		
 	override var tableViewDataSource: TimelineTableDataSource {
 		
@@ -83,63 +83,5 @@ final class MyTweetsContentsController : TimelineContentsController, Notificatio
 			.prefix(maxTimelineRows)
 		
 		dataSource.items = Array(newTweets)
-	}
-}
-
-extension MyTweetsContentsController {
-	
-	final class DataSource: NSObject, TimelineTableDataSource {
-		
-		var lastTweetId: String? = nil
-
-		var items = Array<TimelineTableItem>() {
-			
-			didSet (previousItems) {
-		
-				if let item = items.timelineLatestTweetItem {
-					
-					lastTweetId = item.timelineItemTweetId
-				}
-			}
-		}
-		
-	}
-}
-
-extension MyTweetsContentsController.DataSource {
-
-	func numberOfRows(in tableView: NSTableView) -> Int {
-		
-		return items.count
-	}
-
-	@discardableResult
-	func appendHashtags(hashtags: HashtagSet) -> ProcessExitStatus {
-		
-		let latestHashtags = items.first?.currentHashtags
-		let needAppending = { () -> Bool in
-			
-			switch latestHashtags {
-				
-			case .none:
-				return true
-				
-			case .some(let v):
-				return v != hashtags
-			}
-		}
-		
-		if needAppending() {
-			
-			let item = TimelineHashtagTableCellItem(previousHashtags: latestHashtags, currentHashtags: hashtags)
-			
-			items.insert(item, at: 0)
-			
-			return .passed
-		}
-		else {
-
-			return .aborted(in: -1)
-		}
 	}
 }